Refactor debugger worker
